requirements.txt
# 실행에 필요한 라이브러리 목록 파일
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.3.1
numpy==1.26.3
tqdm==4.66.1